+++ cypress/integration/2.7-discountShipping.spec.js
@@ -26,10 +26,10 @@ describe('Discount Shipping Testcase', () => {
     cy.updateProductQuantity(discountShipping, { quantity: '1' })
   })
 
-  it('Updating Shipping Information', updateRetry(4), () => {
+  it('Updating Shipping Information', updateRetry(3), () => {
     cy.checkForTaxErrors()
     // Update Shipping Section
-    cy.updateShippingInformation({ postalCode })
+    cy.updateShippingInformation({ postalCode, timeout: 8000 })
   })
 
   it('Verify tax and total', updateRetry(3), () => {